Regions & Chunks + Iterators
I like the idea of being able to loop over the data using iterators, it seems like it could be a cool use case to find a certain chunk at a given index, things like that. You could do `BedrockRegion.find(x,y)` for a given Chunk, and it would give you either a Chunk object, or `null`. Something neat like that.