Chunk Data Type Demo
With the new additions I made to NBTify, now `NBTData` objects can be unique in the fact that now you can tell them apart in terms of their save format metadata!

This is super cool because you can use it to tell if two structures are both the same format, and they use the same save format options. If they aren't, then you will get linting errors when you try to assign the wrong format that expects another type.

Say, if you had a Bedrock `level.dat` file, and you converted it to that of a Java `level.dat` file, you might accidentally save the file using little-endian, and with a Bedrock header. Now with this new generic parameter modifier, the two values are different at the runtime level, and the type level. So it will flag you that you are trying to save a little-endian file to a big-endian endpoint, if you set up your types explicitly like that.

This is part of my reason for hoping to get this second generic working, as it can help you treat data structures as blocks of things that aren't the same, and you have to specifically make sure that everything is lined up before you can put the puzzle pieces together again.

You can see here, that you both know the shape of the NBT itself, and what format it should be saved in. So, if any of those are incorrect, you will get a type error for trying to do something that doesn't match the specs you set up. The key here is that you only have to get it right once (kind of, lol), when setting the types up. From thereon out, the type system will help you make sure what you are doing matches what the guidelines you set up were.

EPIC!

*This demo addition was from yesterday, it relied on me finishing up my changes to NBTify and publishing them to npm to be able to use them in this project properly.