Level.dats Schema Start
More work to be done with this one, just getting things started here. Gonna try making a simple converter for Level.dat files between all 3 versions. Hoping to see if this conversion thing is reasonably accomplishable! I think it is with enough work!

I tried spawning in a lot of the different entity kinds in a Bedrock world in hopes of getting their NBT shapes, but I think I'll have to get them another way, as it's hard to manage them all there. Maybe a structure block export could do this nicely, come to think of it!

https://minecraft.fandom.com/wiki/Java_Edition_level_format
https://minecraft.fandom.com/wiki/Player.dat_format