Full ResourceLocation Typings
Got the BlockEntity and TileTick typings from the Wiki. It was a bit different for the Recipe IDs though. As follows here:

Another neat hack to get resource IDs; Rather than getting them from an online resource, I wanted to get all of the IDs for the current version, so I gave all of the recipes to myself in-game; that populates the `level.dat` with all of the in-game available recipes 🤯

```
/recipe give @s *
```

https://minecraft.fandom.com/wiki/Chunk_format#Block_entity_format
https://technical-minecraft.fandom.com/wiki/Tile_Tick_Block
https://minecraft.fandom.com/wiki/Recipe_book#Data_values
https://mcreator.net/wiki/list-vanilla-recipe-registry-names (This was the first option for the Recipe IDs, but I didn't want to worry about the list of them being outdated (almost certain that it was, just because there've been so many new updates the last few years))
https://www.reddit.com/r/Minecraft/comments/6jtgut/can_i_unlock_all_recipes_in_the_crafting_book/ (The method I ended up going with; I read all of the values from `Player.recipeBook.recipes` and `toBeDisplayed`, just to make sure they had the same values)

I wanted to look into if it was possible to merge two union type definitions, like you can with interface merging. Or rather, can you represent a union type using an interface, which then opens the door to being able to use interface merging? I realized that for addons and mods and such, it would be great to provide type definitions for those as well, to ensure you can iterate over the world data with as much high type-safety.

(Made an issue for this, here's some more info and a demo about this idea: #2)

https://betterprogramming.pub/advanced-typescript-how-to-use-interface-inheritance-with-discriminated-unions-dddf77cb3836