ENUMify
Not sure if enums can necessarily help with this (#2), but decided to look into them again since they probably could be helpful for some instances.

Enums vs just a bare type are an improvement no doubt though. I think an enum-like `as const` object could work for this scenario too, I want to look into that to see if it's the more worthwhile option also, since I tend to like that more than how enums limit you to use them.

https://betterprogramming.pub/advanced-typescript-how-to-use-interface-inheritance-with-discriminated-unions-dddf77cb3836