Dimension IDs
Recently rediscovered this repo, and it had some missing information about Dimension IDs for Bedrock. Added those, and got the IDs for Java from in-game in the F3 menu while in each dimension.

https://github.com/Mojang/bedrock-samples

I think it might make sense to organize the files/modules into names that correspond to the Resource Locations that Java organizes it's code into. Or, something similar. I think I'm kind of tending to use these structures for things implicitly already.

https://minecraft.fandom.com/wiki/Resource_location