No-Inheritance Block Entity Definitions
An excerpt from the question discussion in Discord:

For Block Entity references, do you think it should use the mirror of the Block ID if it's the same anyways? Or should it be separate for consistency? I'm thinking separate, as to prevent unexpected correlations between things?

It's kind of confusing, because it kind of goes in 3 stages:
- A Block can be an Item. But,
- An Item isn't necessarily a Block.
- A Block Entity can represent multiple kinds of Blocks, not just a single ID.

So for example, there are individual IDs for each Banner color in Java Edition, but they are only stored under a single `"minecraft:banner"` Block Entity entry. So there isn't a central Block ID with the name `"minecraft:banner"` anymore, because you instead just reference the Block/Item itself using the color.

But for the Block Entities which represent only a single Block, their Block Entity ID is symmetrical to their Block ID, so I'm wondering whether it should just 'inherit' it from that? I kind of feel like no, to help define the degrees of separation for things?

If you change the ID of a block, you wouldn't necessarily want it's Block Entity ID to change with it, so maybe that should be an explicit thing that we should have to update as maintainers, rather than it shadowing that one.

Check out the Banner page on the Wiki for an example of this:
https://minecraft.fandom.com/wiki/Banner

-----

Decided to go with no inheritance because of the separation of concerns/degrees of separation. A point towards this argument is also that Bedrock doesn't follow this standard at all with Block Entities, so that won't be an option for representing things in Bedrock.

While going through the Wiki, turns out there's a *LOT* more info about Bedrock's format than I thought! It's just shown under each feature separately, rather than in all one single place.

I *definitely* will want to make some proper docs/wiki from these types once they are done. It could very much make for a good competator/replacement for the wiki if done right, because the data isn't part of the page like it is on the wiki. I'm thiking something like a Static-Site Generator from these types. I'm pretty sure the wiki doesn't do this, and instead it just has manual data entries with the values for all of the format spec. I want something more easier to manage and work with!!!