No-Enums
Ok, I'll give them that! Enums are definitely really nice to use. They have some drawbacks in my opinion when using them at the type-level though, so I'm gonna go back to my `{} as const` enum-ish union type method.

Funny timing to see @MichiganTypeScript post a new video about this, after having been discussing this recently in the @Minecraft-Manipulator group.

https://www.youtube.com/watch?v=XTXPKbPcvl4 (New enums video)
https://www.youtube.com/watch?v=0fTdCSH_QEU (I think this was the original inspiration for my use of the `{} as const` method.)

This might be a backwards step in relation to #2, but I can come back to that later once the types are more established as a whole. It's still a thing to think about, consider, and brainstorm for though! I definitely want to have that kind of thing be possible.