NBTData Format Generic Revert
From a discussion with HoldYourWaffle, we went back and forth on whether or not it's a good idea to represent `NBTData` format metadata at the type level. In summary, I think Waffle might be right, so I'm going to try doing things without it. I think the main thing here is that it doesn't make sense to work between the formats using `NBTData` wrappers, if you're going between formats. For things like NBTify's CLI and Dovetail, it makes sense to work with `NBTData` wrappers, because you aren't working with specific formats for specific outputs, you're just modifying the current values of the given file you're opening. But for instances like this, where you aren't using the same data as-is, it makes more sense to wrap the data in an `NBTData` wrapper only at the write time. For this view, I think I do see how that could be a better route. I think my hang up here was that I was thinking about conversion code similarly to how Dovetail manages file metadata, while in the case of conversion logic, it doesn't matter as much to bundle that format logic alongside the data itself. It's more the `NBTData.data` part of the conversion that you're worried about here. So, it makes more sense to do that separate to the metadata/format you're writing the file with.

https://github.com/Offroaders123/NBTify/issues/29