Evergreen Aquatic Block IDs
An expansion of the Block IDs from the previous commit, but now with the additional item IDs of LCE Aquatic+ blocks!

Looks like Evergreen's project also uses mainly Flattening+ naming conventions for ID references across the board, so it tends to kind of make it confusing as to which version the enum works with. I may go back and re-1.12ify the names for the enum members, to that of the pre-Flattening appearance. It does make it easier to access members in the conversion stage when both platforms' enum members use the same keys, but this makes it less defined as to where the game versions line up with the types.