Carved Pumpkin Block
Now it makes sense what the duplicate Block IDs for Pumpkin are about, it's because of when in 1.13 it added the 'regular' pumpkin block, which means in LCE's numerical item ID format, the old ID is the carved version, and the new one is the full, un-carved version, the one that now spawns naturally instead of the carved one. Thanks to Evergreen for reminding me about this one, hehe.