Pre-Flattening End Biomes
In the previous commit, I guess I hadn't realized that Bedrock was the only one to use that mix of pre-Flattening and the other name formatting, I thought Java did too, but looks like not.

I'm not sure whether these End biome variants nor the Void biome types are present in LCE, I'm going to look into that further with Evergreen to see whether they should be listed in the enum.

https://minecraft.wiki/w/Java_Edition_data_values/Pre-flattening#Biome_IDs
https://minecraft.wiki/w/The_End_(biome)#Data_values