Java Entity Types
Realized I hadn't actually written the types for these yet, oops! Thought I was further along than that. I also have a lot to work on for BlockStates as well.

Got a bunch done, still about half of the entities left to finish. Already been working on this for a while tonight though, so I'm gonna take a little break, maybe come back to work on it again.

https://minecraft.wiki/w/Entity
https://minecraft.wiki/w/Bedrock_Edition_level_format/Entity_format (use for next time)