Complete Java Mob Typings!!!
Wow, I spent literally all day on this, took a very long time. It's all worth it though, everything is all in one place now :D

I made it so the various modifying traits no longer inherit from `EntityLike`, and instead you pick and choose which ones you would like for that shape. This is more composable I think, and yuck inheritance.

I still have to make better `HorseLike` and `LlamaLike` traits, as well as make `EntityLike.id` a generic, rather than needing to define it on every single instance itself. I want it to let you know when you extend `EntityLike`, but you forgot to define an ID. That's one of the most crucial parts of it, so I want to have that easily definable, as well as validated at the type level to ensure it's defined when you make the new shape.

https://minecraft.wiki/w/Mob#List_of_mobs
https://minecraft.wiki/w/Mob#Platform_availability (This was the table I scrolled down to get to all of the various articles quickly)

I still need to make the TypedArray tags generic as well, so I can make improved types like `UUID`, which would be of type `IntArrayTag<[number, number, number, number]>`, and also a position type that would be `IntArrayTag<[number, number, number]>`. Having those are going to be crucial for converting between Java, Bedrock, LCE, and anything under the sun.

Listened to sooo much music, don't know I can even remember it all to sum it up. Need to go to bed, then have a good breakfast, and I want to get a nice run in tomorrow. My body clock is all messed up now from working late the last few nights, and staying up this late tonight as well.