EntityLike ID Generic
Wow, that's more than a hundred less lines from the file! That wasn't the goal, but always nice I guess. It does make the interface's definition line a bit longer, but I like that it's strongly-typed now. That was my main goal in doing this. You aren't able to `extends EntityLike` without providing an ID string for the entity.

Ride the Lightning, yeah Cliff Burton was a badass.