Bee Nest / Beehive Actor Overlap
Turns out that Bedrock does actually use the same Savegame ID kind for each of the two blocks (so I mostly have to revert the previous commit). This is a bit hectic because you have to check the SubChunk data to see what the actual block ID is, as now we know they aren't the same ID between the Actor and the Block State.

I tested this out through loading a demo world I made with my Bedrock-LevelDB project, and read through all of the Block Entities in the world to see what their entries are. Indeed, they were all with ID `Beehive`, even though there were 11 total, 5 were Bee Nests, 6 were Beehives.

Another thing I found is that the `Entity.internalComponents.EntityStorageKeyComponent.StorageKey` is of type `StringTag`, but it's characters are not valid UTF-8, so I think it breaks parsing it as SNBT. I think this might be counter-able with the use of implementing MUTF-8, since I think unsupported characters get escaped when parsed as a string. I will have to look into this one. I'd save a demo of this to NBTify so I can validate that what I'm reading is what I'm writing, but I'm not sure I'm currently reading the bytes correctly as they are either, so I'll make sure they are being read right before trying to test against them.

https://minecraft.wiki/w/Bedrock_Edition_level_format/Block_entity_format#Beehive_and_Bee_Nest
https://minecraft.wiki/w/Beehive#ID
https://minecraft.wiki/w/Bee_Nest#ID