Platform Barrel Exports
Going to start demoing use of the current types a bit more, to see how they work together in terms of how I need to author these as modules themselves. Should I add platform prefixes to the names of everything, or should I just scope them to a namespace, or use them with `import * as Platform` instead? Most of the time it's just types anyways, so I think that's alright to do it like that. I also would want to figure out another possible way of doing it though.

Too bad you can't do `export * as Platform` or something like that. By that, I mean if you were make everything prefixed with that name. Like `BlockEntity` turns into `PlatformBlockEntity`. `Platform.BlockEntity` works too, but I don't want to have to import all of everything for the platform, the only reason for needing to do that with this scenario is because each of the shapes are named essentially the same thing, unfortunately. So I either need to rename things to be platform-unique, or to also possible `export { Thing as PlatformThing }`, which is a bit hectic as well.

I still want to add all of my types from Bedrock-LevelDB to this project too! Making lots of progress on Bedrock documentation over there. Still a bit more mysterious files to figure out, but getting closer and closer with every little step :P

Wanting to fully step in and figure out LCE chunks too, that's bugging me, me not being able to figure out how they work.