Bedrock Entity Types (All!)
Another heavily productive day. Trying to get faster at just pulling these off all in one go (not actually in a single sitting, but yeah. I did this in about two/three I think), maybe a feature a day. Did it during the daytime, rather than at night, really had to fix my sleep schedule a bit. Now I am feeling better back on track with that.

TS is foogin gold man, I love it. I can't wait till I get the Rust hype too, I think that just hasn't come my way yet. I want to like it, and I know I will eventually, but I don't think I'm quite ready for it yet, still lots I want to explore with this, TS and JS :)

https://minecraft.wiki/w/Bedrock_Edition_level_format/Entity_format
https://minecraft.wiki/w/Zombie_Villager#Entity_data
https://minecraft.wiki/w/Evoker#ID
https://minecraft.wiki/w/Effect#Effect_list
https://minecraft.wiki/w/Arrow#Metadata
https://minecraft.wiki/w/Punch
https://minecraft.wiki/w/Bedrock_Edition_level_format#NBT_Structure
https://minecraft.wiki/w/Splash_Potion#Entity
https://minecraft.wiki/w/Bedrock_Edition_data_values#Effect_IDs

Gotta say, the Wiki is the foogin best as well, insane kudos to everyone who has laid the groundwork for this over all of these years, the documentation here is great! Considering how much of a "mess" Bedrock's NBT (and world save format) kind of is (NBT on all platforms tbh), they have a lot of great info here, at least something to start with.

Thankfully there are actually a good deal of similarities in the rough of inconsistencies, a lot of the older features use the same NBT shapes and properties, which will make them easy to convert between. Anything added recently, as of more modern world format adjustments, or new features in general, it doesn't seem like they plan for implementation parity too well, which is unfortunate. It does look like 1.14+ villagers may have the same shapes between Bedrock and Java (if my exploded brain was reading and remembering my code right), so that's nice. Things like that, are similar. It's the block, block entity, and effect handling properties which go all over the place, anything to do with mob or block variants or states, that's going to be the largest headache out of all of this. The main thing here is that I just need to plainly document all of the inconsistencies in how they are represented through their medium, as an implementation itself, then I can just write a streamlined intermediate format that represents all of these in a common ground, which is easily understandable to see. It will probably lean towards the Java way of doing things, in some sense of things. Especially since LCE is part of the picture too, and that tends to lean Java for most things.

Another interesting thing is the use of IDs between Java and Bedrock. Java uses `IntArrayTag<[number, number, number, number]>`, while Bedrock uses just a plain `LongTag`. Can these both losslessly represent the same UUID, in either representation? Or are they using different UUID/ID schemes? If they are the same, then this is super extra awesome, because then Entities can keep their references the same across the whole world, and the only thing that will need mapping is Player IDs, but depends if they are MS account driven, vs Mojang UUID driven. Wait, with Mojang accounts being merged, are the original UUIDs system for Java still in use? And if not, does that mean that players have the same UUID for Bedrock and Java? That sounds heavily too hopeful for me, but it's a thought to think at least. Either way, the Entity one would be cool, so yeah. Hope that works out.

I'm going to think more about exploring the wild west of LCE again, that is the furthest progress of documentation out of the three that I currently have. I've done a good deal on the file formats themselves, but I really need to get more of the types setup properly. This project seems to be getting vastly closer at the same speed that it's getting more apparent to me just how large of an undertaking it is. The more work I do, the more work it seems like there is to do. I know that's how things go, but haha, it really do be like that. I'm just keeping going with it while the times of working on it are fun. Just like how playing MC itself became less of a fun thing for me over the last few years, this project most certainly will be as well. While the iron is hot, I gotta get stuff done!

Was listening to this for the last few hours of finishing up these changes just now, I randomly remember it's on, then forget again. Very great sound for ambience in the room while working. Almost got to the 2 hour mark, geeez. I have to go to bed NOW. Already too late. Ok good night, cya friends :p
https://www.youtube.com/watch?v=wPhefMiF5Zw

:wq

lol I wrote this partially in GitHub Desktop, then came back and amended a heck bunch of it from VSCode.