Registry ID & Names
Not fully typed accordingly yet, but it is working value-wise at least. When you `switch` on `entry.name` or `entry.id`, the vice-versa for one of those two isn't persisted, it doesn't get narrowed. How can I both derive a discriminant generic from an enum, and also make sure to restrict that generic to only allow enum-like structures in?