Resource String Autocomplete!
By chance, rediscovered how to list enum members as a string for a given type! You just wrap the enum in a template literal expression, and it will evaluate the values for each of those.

Is there a way to make that same transform, but for a number-based enum as well? I want to allow autocomplete for number-based enums as well, so you don't have to specifically use the enum itself to see the various options available.

Now the player's `identifier` key in the Java types is marked as `undefined` again, since omitting it from the type altogether means you can't validate it's value in an equals check safely, at least in terms of TS at the type level. In order to safely do checks like that, all of the shapes in the union type have to have the same member that you want to validate with.

Looked more into the item formats for each platform. I still have to add types for all of the things there.

https://minecraft.wiki/w/Bedrock_Edition_level_format/Item_format
https://minecraft.wiki/w/Item_format

SYL Alien, followed by MK Wooden Smoke, is making me feel better :)