Pi Entity ID Usage
Brought this over from Java's types.

Unfortunately I can't use the `switch (entity.id){}` trick for `IntTag<>` wrapped keys, I'm sad about this. How can I deduce the result of `valueOf()` so it can properly scope if checks to that of the entity's ID property? It works for strings, but how can I wrap that association to the object itself. This reminds me of the same kinds of issues with Web Components, where you have to make the outside world think your custom wrapper can be used like a plain thing, like in regards to things like form controls for example.

https://github.com/BJTMastermind/MCPI-Parser
https://stackoverflow.com/questions/67863114/is-there-a-way-to-have-a-switch-statement-for-an-object-with-cases-that-validate (funny that the question is also about a Minecraft thing! :D)