LCE Item IDs (Updated)
Thanks again to @PhoenixARC for helping find these from the decompiled source code from the game. I kept the ID for the Shield around, even though the new typings reference didn't have those as a definition, because I think PS4 edition may use that ID since it supports the Shield item. PS4 is the only LCE platform to have Shield support, because it's the only one that received Village & Pillage.