LCE Enchantment IDs (Updated)
Got some more updated values from PhoenixARC, my previous ones weren't listed correctly compared to these new ones.

Found another fellow early/solo band making songs as well, like me and Ethapher do. It's great!
https://www.youtube.com/watch?v=hamfbfLdVwI

Listenede to Normalizer 2 twice through today. Going to try my own take at Marco Minnemann's project, to write music around his 50 minute drum solo. I already know it fairly well by way of the other projects that have already built on top of it (namely Evidence of Humanity). I want to get to know it a little more solid on it's own though, I think things will come more naturally to me easier during the writing process, if I'm more familiar with the changes, and the general different moods/tempos that happen throughout the song. If I can discern what each "section" of sorts does for me, then I think I can internalize that and discover how that sounds when I write it as music. I think I possibly do that on my own already, for my own music. The thing here though is that I don't have the pre-association with knowing the groundwork for what makes each part of the drum solo (in a creative sense), so I have to understand where each part kind of leans towards, so I can feel relatively comfortable in making music for it. Not that I want to completely make it perfect or something, but I want to try to not just make random noises on the top of it, since there is an underlying rhythm and pattern to it, I just have to get to know that. I already discovered one of the parts is in 9, which I hadn't figured out yet. I think the challenging parts about Marco's solos are the use of (backbeats?) and ghost notes. Let alone it's usually in a different signature, but it's not as bad to deduce that because he (like all of us) tends to go in a specific way, and I think I understand where he's coming from with those time changes, now. Now it's just the part of getting familiar with the specific paths he took here to tie these concepts together to make the solid piece that turned out to be the whole drum solo. I can already hear a few different genre spots that I can play for in the beginning, as well as the end. The middle parts are where I don't know how to write for yet.