Java Effect IDs (Updated)
Re-remembered that this was kind of an awkward union type of values before, since effects in Java are both defined with either the ID, or the resource identifier. This is about to be less of a problem because of the upcoming changes to NBT in the recent snapshots this week. It's still important to have these documented for versions that still had this though, since hopefully I want to have everything supported to where going between any given version should be fairly easy (in a perfect world, lol. I know there will be tradeoffs. It only helps to have everything documented though).

https://minecraft.wiki/w/Effect#Effect_list

This was the little script I wrote to generate an object from all of the values in the Effect definitions table.

```ts
Object.fromEntries([...$0.querySelectorAll("tr")].map(tr => [tr.querySelector("td:nth-child(3) code")?.innerText,Number(tr.querySelector("td:nth-child(6)")?.innerText)]))
```

Here's a huuge bunch of the changes for the snapshots, in relation to NBT. Region-Types is about to get a lot of work for it. Hey, in the name of a better Minecraft future though! This was the same thing with 1.13 as well, I'd rather have that nice setup for things, than not.
https://www.youtube.com/watch?v=iY9OHAd4Aco