LCE PS4 Effect Possibilities
Do any of these Effects exist on LCE? Maybe PS4, since that got Village & Pillage.

https://www.youtube.com/watch?v=IMwNGnyq0lU
Dang, LCE with Jappa textures really did look great. Sad that LCE isn't still developed nowadays! I wonder if it would be possible way down the line, to reverse-engineer some of it's code, to build a native version for other devices, and have LCE anywhere. Would be great with a controller too. Or I could just learn to emulate it. I would like to see if something to the equivalent of MCPi Revival was possible too, just to a little bigger scale.