Bedrock Block State Typings
The world format from my current knowledge still uses raw byte data to store the block state variants, but there does appear to be a spec for these variants as NBT compound tags as well, which is what things like Structure Block exports and things similar, use. I think it's what the game might use internally too, say like if Bedrock were to expose the runtime NBT of things to the user via commands. It sounds like this is what the server protocol might be starting to use as well, so there's only hope that the world format itself will finally use these soon now as well!

Thanks to @PMK744 for the help in getting access to the list for these!

This is the tool I made to convert that list into these type definitions.

https://github.com/Offroaders123/State-Generator