PascalCase Renaming
Now the entire setup from these two commits is done automatically by that builder script! Might as well not have to do that same text manipulation over and over if I already am close enough to being able to automate that part.

I need to solve the naming issues for differentiating between Block States, Block Entities, and things like that. Right now Region-Types just uses the same names in each file on their own, but this doesn't collide nicely when exporting them all from the main package as a whole. I have to find a way to manage that for the naming for each platform as well, say because each edition has their own `Cow` definition for example.