Accidental Block State Field Coalescing
Big oops! Now things should be fixed now. Now it makes sense why so many tags were inconsistently typed as well. I thought it was just Bedrock having a Bedrock moment, so I didn't think too too much about it until I saw that the wall typings looked off for sure.

https://github.com/Offroaders123/State-Generator/commit/ab06ca26208f69f61e72e43e7dcb46c513f7f618

I fixed this with these changes by way of adding the stack of commits/fixes to a new side branch, and I was going to try to merge-commit them back into the main branch. That didn't fully workout nicely though, so I tried to rebase them into the history of the current commits on the main branch (while maintaining the original errored-history, for learning's sake), but that got a bit conspicuous as well, so I rebased the commits for that into this single commit, and rebase-copied this over to this main branch. That's why the changes are nice and smooth. I had to manually add the deduping and pascal naming features to the new source again, the tool doesn't do that for you. That's extra niceties that I've added on top of what it's output for us.

TypeDoc for Minecraft NBT types :D
That would work excellently for things like the wiki, honestly feels more maintainable as well, since it's code-first driven. I still have to go back and add JSDoc to all of these platform typings.