Java 1.20 Folder
The typings for modern Java Edition currently are up to about v1.20. I am looking into completing the typings for all of the versions, the best I can to some extent. I just found out that much of the changes and documentation over time on the wiki are much more accessible than I originally thought, which is outstanding! Now I feel safe in doing this completely right. Things don't need to be completely absolutely perfect at first, I need to get a little bit here and there to get started on it.

Eventually I kind of want to do all of this in Rust now, but that is a bit of a ways away still. I don't want to break parsing things for the sake of type safety at runtime. That should be an update to things later on once it's needed. Right now it's more important to "move fast" to some extent (keeping quality and accuracy of data types). This setup is plenty good enough to at least document all of this more thoroughly to start with. Once any issues or missed concepts show up, then we can fix them. Having some sort of typings is better than none. Versions are always static once released too, so it's not like we have to worry about 1.8 changing for example. Same with all of the other ones.