Java Alpha Item IDs
The wiki is awesome. So glad all of this is archived and still available for us to access. Diffing works too, that is making things magnitudes simpler, big thanks.

https://minecraft.wiki/w/Java_Edition_data_values?oldid=38061

I am using a "Beta"-dated version of the docs, because it seems to be the latest version of this page before any Beta "features" were introduced. It looks like Cake was the first thing to be added, so I did the newest log up to that change. I found this by lining up the dates with when Alpha versions came out, and the first Beta version did. At the time of Beta having been released, this page is still accurate to the information about Alpha back at that point.

https://minecraft.wiki/w/Java_Edition_version_history#Beta
https://minecraft.wiki/w/Java_Edition_Beta_1.2

I wonder if people have archived the wiki into a Git repo too. That seems like it would be crucial, if it hasn't been done already. Having that could very much help for tracking changes of data types over time too, rather than having to manually check this on one's own. That would be the ideal scenario in concept though. This is working out great still so far.