Java Alpha Block IDs
Okay, so cloth was in Indev, and it was removed as part of Alpha. Good to know. Interesting that wool wasn't re-added yet either, so a neat migration in that sense isn't totally for certain, our mappings could allow for nice ways to handle this though.

Great to see much of these types are still the same to other editions and versions, to some extent. At least for LCE, most of these have lined up very well.